Usuarios 1.- Actualización de los modelos - Agrego importaciones de los demas modelos para que existe consistencia. 2.- Creación del esquema de usuarios 3.- Creación de 2 routes - autenticación de usuarios (auth_usuarios) - creacion de usuarios (usuarios)

+++ db/models/materias.py
@@ -4,10 +4,10 @@ from ..conexion import Base
 
 class Materias(Base):
     __table__ = "materias"
-    id = Column(Integer(11), primary_key=True, index=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     nombre = Column(String(20), nullable=False)
-    codigo = Column(Integer(11), nullable=False)
+    codigo = Column(Integer, nullable=False)
     descripcion = Column(String(200), nullable=False)
-    creditos = Column(Integer(5), nullable=False)
+    creditos = Column(Integer, nullable=False)
 
     matriculas = relationship("Matriculas", back_populates="materias")
